requirements.txt
Django==1.6.2
List==1.3.0
Pillow==2.3.0
South==0.8.4
Werkzeug==0.9.4
argparse==1.2.1
beautifulsoup4==4.3.2
bleach==1.4
boto==2.25.0
dj-database-url==0.2.2
dj-static==0.0.5
django-extensions==1.3.3
django-filter==0.7
django-nose==1.2
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.12
factory-boy==2.3.1
gunicorn==18.0
html5lib==1.0b3
ipython==1.1.0
mock==1.0.1
nose==1.3.0
pip-tools==0.3.4
psycopg2==2.5.2
pystache==0.5.3
pytz==2013.9
setuptools==1.0
six==1.5.2
static==1.0.2
unittest2==0.5.1
wsgiref==0.1.2
